hardening(types): declare(strict_types=1) généralisé sur les points d'entrée (368 fichiers) + casts float défensifs sur les built-ins numériques
- Ajoute declare(strict_types=1) aux pages/entrypoints (modules, admin, accueil,
  login, API, rgpd, parametres, tenant, tutorat…) après les services (vagues 1-2).
- Neutralise les TypeError latents des built-ins numériques sur données PDO
  (chaînes) : number_format/round/floor/ceil/abs/floatval reçoivent désormais
  (float) (…) — 164 sites. Corrige le seul TypeError observé au crawl
  (facturation/factures.php : number_format(string)).
- Préserve les fins de ligne d'origine (CRLF conservés là où HEAD les avait).

Vérifs : phpunit 125/125 ; crawl 5 rôles × 208 routes = 0×500 / 0 TypeError ;
pages à calculs (bulletins/factures/reporting/inventaire/stats/notes) = 0 TypeError ;
saisie de notes bout-en-bout (prof → 302, note insérée) = 0 TypeError ;
lint OK sur les 401 fichiers ; aucun built-in string recevant un int détecté.

+++ API/Admin/AdminCrudPage.php
@@ -421,7 +421,7 @@ class AdminCrudPage
 
     private function renderPagination(): void
     {
-        $totalPages = max(1, (int) ceil($this->totalItems / $this->perPage));
+        $totalPages = max(1, (int) ceil((float) ($this->totalItems / $this->perPage)));
         if ($totalPages <= 1) return;
 
         $params = $_GET;
